http://jira.bytatech.com:8080/browse/IN-7692

+++ MockExam/src/main/java/com/lxisoft/MockExam/repository/OptionsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import com.lxisoft.MockExam.model.Options;
 
 @EnableJpaRepositories
-public interface OptionsRepository extends JpaRepository<Options,Long> {
+public interface OptionsRepository extends JpaRepository<Options,Integer> {
 
 }
